In order to improve maintainability and code organization, reconstruct views.py into multiple specific modules.
matcher/views/

├── __init__.py # Import all views to maintain backward compatibility

├── auth_views.py # Authentication-related views (OAuth, login, logout)

├── main_views.py # Home page views and work matching logic

├── profile_views.py # User profile management

├── job_views.py # Job detail page

├── application_views.py # Job application-related (application letter, resume customization, application management)

└── experience_views.py # Work experience management
