+++ job_hunting_project/urls.py
@@ -21,7 +21,7 @@ from django.conf import settings
 from django.conf.urls.static import static
 from django.conf.urls.i18n import i18n_patterns
 from django.contrib.auth import views as auth_views
-from matcher import views as matcher_views  # Import matcher views directly
+from matcher.views.auth_views import google_login, google_callback, process_oauth_tokens  # Import specific auth views
 
 urlpatterns = [
     path('i18n/', include('django.conf.urls.i18n')),
@@ -30,9 +30,9 @@ urlpatterns = [
     path('login/', auth_views.LoginView.as_view(template_name='matcher/login.html'), name='login'),
     path('logout/', auth_views.LogoutView.as_view(next_page='matcher:main_page'), name='logout'),
     # OAuth routes should not be internationalized to avoid session issues
-    path('auth/login/google/', matcher_views.google_login, name='google_login'),
-    path('auth/callback/', matcher_views.google_callback, name='google_callback'),
-    path('auth/process-oauth-tokens/', matcher_views.process_oauth_tokens, name='process_oauth_tokens'),
+    path('auth/login/google/', google_login, name='google_login'),
+    path('auth/callback/', google_callback, name='google_callback'),
+    path('auth/process-oauth-tokens/', process_oauth_tokens, name='process_oauth_tokens'),
 ]
 
 urlpatterns += i18n_patterns(
