+++ matcher/views/README.md
@@ -1,0 +1,91 @@
+# Views 重构说明
+
+## 重构概述
+
+原始的 `views.py` 文件包含了超过 800 行代码，为了提高可维护性和代码组织，我们将其重构为多个专门的模块。
+
+## 新的文件结构
+
+```
+matcher/views/
+├── __init__.py              # 导入所有视图以保持向后兼容性
+├── auth_views.py            # 认证相关视图（OAuth、登录、登出）
+├── main_views.py            # 主页视图和工作匹配逻辑
+├── profile_views.py         # 用户个人资料管理
+├── job_views.py             # 工作详情页面
+├── application_views.py     # 求职申请相关（求职信、简历定制、申请管理）
+└── experience_views.py      # 工作经验管理
+```
+
+## 各模块职责
+
+### 1. `auth_views.py` - 认证管理
+- `google_login()` - 启动 Google OAuth 登录
+- `google_callback()` - 处理 OAuth 回调
+- `logout_view()` - 用户登出
+- `api_check_auth()` - API 认证状态检查
+- `process_oauth_tokens()` - 处理客户端 OAuth tokens
+- 辅助函数：`get_current_user_info()`, `logout_user()`
+
+### 2. `main_views.py` - 主页和匹配逻辑
+- `main_page()` - 主页视图，处理工作匹配
+- 内部辅助函数：
+  - `_handle_oauth_callback_on_main_page()` - 处理主页上的 OAuth 回调
+  - `_handle_job_matching_post()` - 处理工作匹配 POST 请求
+  - `_save_job_matches_to_db()` - 保存匹配结果到数据库
+  - `_handle_session_view_get()` - 处理查看特定匹配会话
+  - `_process_matched_jobs_for_session()` - 处理匹配会话中的工作列表
+
+### 3. `profile_views.py` - 用户资料
+- `profile_page()` - 用户个人资料页面
+  - CV 文本管理
+  - PDF 文件上传和文本提取
+  - 用户偏好设置
+
+### 4. `job_views.py` - 工作详情
+- `job_detail_page()` - 工作详情页面
+  - 显示工作信息
+  - 匹配分析和建议
+  - 异常分析显示
+  - 申请状态管理
+
+### 5. `application_views.py` - 申请管理
+- `generate_cover_letter_page()` - 生成求职信
+- `generate_custom_resume_page()` - 生成定制简历
+- `download_custom_resume()` - 下载简历 PDF
+- `my_applications_page()` - 我的申请管理
+- `update_job_application_status()` - 更新申请状态
+
+### 6. `experience_views.py` - 经验管理
+- `experience_list()` - 工作经验列表
+- `experience_delete()` - 删除工作经验
+- `experience_completed_callback()` - N8n 完成回调
+
+## 向后兼容性
+
+为了确保不破坏现有代码：
+
+1. **原始 `views.py`** 现在只是一个简单的导入文件，从新的子模块导入所有视图
+2. **URL 配置** 无需更改，所有视图仍然可以通过 `views.function_name` 访问
+3. **导入语句** 在其他文件中的 `from matcher import views` 仍然有效
+
+## 重构的好处
+
+1. **代码组织** - 相关功能归类到同一模块
+2. **可维护性** - 更小的文件更容易理解和修改
+3. **测试** - 可以针对特定功能模块进行单元测试
+4. **团队协作** - 不同开发者可以专注于不同的功能模块
+5. **分离关注点** - 每个模块有明确的职责
+
+## 未来扩展建议
+
+1. **进一步优化** - 可以考虑将一些大的视图函数进一步拆分
+2. **业务逻辑分离** - 考虑创建 service 层来处理复杂的业务逻辑
+3. **API 视图** - 如果需要 API 端点，可以创建专门的 `api_views.py`
+4. **缓存策略** - 为频繁访问的数据添加缓存层
+
+## 注意事项
+
+- 确保所有新的视图模块都在 `__init__.py` 中正确导入
+- 主项目的 URL 配置已更新为直接导入认证视图
+- 所有的导入和依赖关系已经正确配置
