+++ requirements.txt
@@ -1,7 +1,7 @@
 Django~=5.0.4
-google-generativeai~=0.5.0 # Or the latest version available
+google-generativeai~=0.5.0
 supabase>=2.0,<3.0
 python-dotenv
 PyMuPDF~=1.24.1
 django-cors-headers~=4.3.1
-reportlab~=4.0.0 
+reportlab~=4.0.0
